Start incorporating devcards
To make the most out of devcards, a lot of components are going to
need to be rewritten so they are easily testable. This mostly means
building complicated components out of a single container that sorts
out handlers and subscriptions, which it then passes to smaller, easily
tested, static components.